Fix usigned integer overflow resulting in completely wrong vote_weight values

+++ contracts/stakevote/stakevote.hpp
@@ -17,7 +17,7 @@ CONTRACT stakevote : public contract {
     using config_container = eosio::singleton<"config"_n, config_item>;
     struct [[eosio::table("config"), eosio::contract("stakevote")]] config_item {
         // time multiplier is measured in 10^-8 1 == 0.00000001
-        uint16_t time_multiplier = (uint16_t)100000000;
+        uint32_t time_multiplier = (uint32_t)100000000;
 
         static config_item get_current_configs(eosio::name account, eosio::name scope) {
             check(config_container(account, scope.value).exists(), "Stake config not set.");
